Added file upload and download

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const { validationResult } = require('express-validator/check');
 
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -15,10 +16,25 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const description = req.body.description;
   const price = req.body.price;
-  const userId = req.user._id;
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: 'Attached file is not an image.',
+      validationErrors: []
+    });
+  }
 
   const errors = validationResult(req);
 
@@ -30,7 +46,6 @@ exports.postAddProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description
       },
@@ -39,12 +54,14 @@ exports.postAddProduct = (req, res, next) => {
     });
   }
 
+  const imageUrl = image.path;
+
   const product = new Product({
     title: title,
     price: price,
     description: description,
     imageUrl: imageUrl,
-    userId: userId
+    userId: req.user
   });
 
   product
@@ -94,7 +111,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const productId = req.body.productId;
   const updatedTitle = req.body.title;
-  const updatedImageUrl = req.body.imageUrl;
+  const updatedImage = req.file;
   const updatedDescription = req.body.description;
   const updatedPrice = req.body.price;
 
@@ -108,7 +125,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDescription
       },
@@ -127,7 +143,11 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDescription;
-      product.imageUrl = updatedImageUrl;
+      if (updatedImage) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = updatedImage.path;
+      }
+      
       return product
         .save()
         .then(() => res.redirect('/admin/products'))
@@ -166,9 +186,15 @@ exports.postDeleteProduct = (req, res, next) => {
   const productId = req.body.productId;
 
   Product
-    .deleteOne({
-      _id: productId,
-      userId: req.user._id
+    .findById(productId)
+    .then(product => {
+      if (!product) return next(new Error('Product not found.'));
+      fileHelper.deleteFile(product.imageUrl);
+      return Product
+        .deleteOne({
+          _id: productId,
+          userId: req.user._id
+        });
     })
     .then(() => res.redirect('/admin/products'))
     .catch(err => {
